refactor(chunk-store): Phase 1b — move WAL bodies to chunk_wal.c
Moves csReplayWal and its replay/reload-state helpers (~280 lines)
out of chunk_store.c and into chunk_wal.c, completing the
chunk-store decomposition started in Phases 1–5.

Public-API additions in chunk_wal.h:
  - ChunkStoreReplayState, ChunkStoreReloadState struct defs
  - csReplayWal, csCaptureReloadState, csFreeReloadState,
    csAdoptOpenedStoreState (called from chunk_store.c)

Replay-state capture/restore/release/rollback stay file-local
in chunk_wal.c since csReplayWal is the only caller.

To support the move, this also exposes (un-statics + declares)
helpers that csReplayWal/csAdoptOpenedStoreState call across the
subsystem boundary:
  - refs:    csCaptureSavedRefsState, csRestoreSavedRefsState,
             csFreeSavedRefsState, csReplaceRefsStateFromBlob,
             csFreeRefsState, csAdoptRefsState,
             csEnsureDefaultBranch
             (plus SavedRefsState struct moved to chunk_refs.h)
  - staging: csGrowPending, csPendHTClear, csRecentHTClear
  - index:   csSearchIndex, csMergeIndex, csReleaseIndexBuf
  - file:    csCloseFile

Verified:
  - make doltlite-lib + make doltlite (clean builds)
  - make lint (layering check passes)
  - Smoke: CREATE TABLE + dolt_commit + reopen, dolt_log shows 2 entries
  - WAL regression tests:
      truncated_wal_rejected            (6/6)
      wal_offset_corruption             (5/5)
      reload_refs_transactional         (13/13)
      refresh_refs_corruption_preserves_state (13/13)
