+++ src/chunk_file.h
@@ -23,4 +23,6 @@ i64 chunkFileGetSize(const ChunkFile *cf);
 void chunkFileSetHandle(ChunkFile *cf, sqlite3_file *pFile);
 void chunkFileSetSize(ChunkFile *cf, i64 nSize);
 
+void csCloseFile(sqlite3_file *pFile);
+
 #endif
